Moved screenshot inside github.
I did this so I am not depending on Evernote to keep the screenshot around.
Also, this makes it nicer to be able to reference in the README.md later.
Beyond that I removed the download zip and tar options because those don't
really make sense given how the plugin should be installed these days.